건강기록 그래프 디자인 수정 전:
