+++ src/views/health/HlogDetail.vue
@@ -2,8 +2,7 @@
 import { onMounted, reactive, ref } from "vue";
 import moodLevels from "@/assets/health/moodLevels.json";
 import sleepQualitys from "@/assets/health/sleepQualitys.json";
-import HealthCart from "@/components/health/HealthChart.vue";
-import HChart from "@/components/health/HChart.vue";
+import HealthChart from "@/components/health/HealthChart.vue";
 import { useHealthStore } from "@/stores/healthStore";
 import { useRoute } from "vue-router";
 import { getHlog, deleteHlog } from "@/services/health/hlogService";
@@ -117,7 +116,7 @@ const deleteLog = async () => {
       </div>
     </v-item-group>
     <!-- 통계 그래프 -->
-    <HChart
+    <HealthChart
       :selectedDate="state.hlog.healthlogDatetime"
       :selectedField="selectedField"
       :fields="fields"
